requirements.txt
llama-index
llama-index-vector-stores-mongodb
llama-index-embeddings-openai
pymongo
python-dotenv
pdfplumber
google-colab